Reshuffle all the directories and file names.
(1) Remove all unnecessary leading underscores.

The leading underscore is necessary on the initial _js directory to
prevent Jekyll from outputting the Javascript files into production.

A config exclude does not work because Jekyll only understands how
exclude top-level directories.

(2) Rename all components to CamelCase. This follows the AirBNB style
guidelines for React components.

(3) Move components into _js/components/{panel}/ subdirectories.
Questions are now located along their panel. It may be worth creating
another level for each section but seems overkill for now.
