+++ _health-care/_js/components/IntroductionPanel.jsx
